requirements.txt
azure-ai-textanalytics 
pandas